# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet Management APIs",
        "description": "API for managing wallets and transfers with double-entry bookkeeping",
        "contact": {
            "email": "samspike46@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/api/v1/transfers": {
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Initiate transfer between wallets",
                "description": "Send money from one user's wallet to another. Must validate: Sender != receiver, Sufficient balance, Atomicity (both debit & credit must succeed).",
                "operationId": "initiateTransfer",
                "requestBody": {
                    "description": "Transfer initiation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sender_wallet_id",
                                    "receiver_wallet_id",
                                    "amount"
                                ],
                                "properties": {
                                    "sender_wallet_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "receiver_wallet_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1,
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Payment for services",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transfer initiated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient balance or validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
        "/api/v1/transfers/{id}": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "View transfer details",
                "description": "Fetch details of a specific transfer including sender, receiver, timestamps, and status.",
                "operationId": "getTransfer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transfer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallets/{walletId}/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "View all transfers for a wallet",
                "description": "List all incoming and outgoing transfers for a specific wallet.",
                "operationId": "getWalletTransfers",
                "parameters": [
                    {
                        "name": "walletId",
                        "in": "path",
                        "description": "Wallet ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by transfer type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "incoming",
                                "outgoing",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet transfers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "description": "Standard API response format",
                "required": [
                    "status_code",
                    "message"
                ],
                "properties": {
                    "status_code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "description": "Response message",
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "description": "Response data",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "errors": {
                        "description": "Array of error messages",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Error response format",
                "required": [
                    "status_code",
                    "message"
                ],
                "properties": {
                    "status_code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "description": "Array of validation errors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "The amount field is required."
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "tokenAuth": {
                "type": "apiKey",
                "description": "Use token: VG@123 in header",
                "name": "token",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "tokenAuth": []
        }
    ],
...
}